Hw04 auth

+++ models/contacts/removeContact.js
@@ -4,8 +4,8 @@ const DB = require('../../config/db.js');
 // const db = new DB('../db/contacts.json');;
 
 
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({ _id: contactId });
+const removeContact = async (contactId, user) => {
+  const result = await Contact.findOneAndRemove({ _id: contactId, owner: user.id });
   return result
 }
